refactor(scope): extract shared registry-artifact sync helper
Deduplicate PROJECT-DEFINITION and specification sync logic into
registry-artifact-sync.ts to keep branch coverage above threshold
while preserving parallel public sync entry points.
